requirements.txt
asgiref==3.3.1
Django==3.1.8
pytz==2020.5
sqlparse==0.4.1
django-mptt==0.14.0
django-js-asset==2.0.0